test(v3): move vault test helpers out of conftest
`from conftest import ...` is ambiguous the moment another test directory adds
a conftest of its own — which is likely while several SPECs land in parallel.
Constants and helpers now live in `vault_helpers.py`; conftest keeps only the
fixtures pytest collects by itself.
