+++ v3/server/tests/vault/conftest.py
@@ -1,40 +1,19 @@
-"""Shared fixtures for the vault-engine tests."""
+"""Fixtures for the vault-engine tests. Plain helpers live in vault_helpers."""
 
 from __future__ import annotations
 
-from collections.abc import Awaitable, Callable
 from pathlib import Path
 
 import pytest
+from vault_helpers import TEST_POLICY, EngineFactory
 
-from palaia_hub.vault import Attribution, EventBus, GitPolicy, VaultEngine
-
-#: Git policy for tests: repack early, never in the background (deterministic
-#: sizes), and treat locks as stale quickly so crash recovery is testable
-#: without sleeping for the production threshold.
-TEST_POLICY = GitPolicy(
-    gc_auto=64,
-    gc_auto_pack_limit=4,
-    gc_detach=False,
-    gc_commit_interval=64,
-    stale_lock_after=0.25,
-)
-
-TEST_ATTRIBUTION = Attribution(
-    agent="curator",
-    client="claude-code",
-    provider="anthropic",
-    session="s-42",
-)
+from palaia_hub.vault import EventBus, GitPolicy, VaultEngine
 
 
 @pytest.fixture
 def anyio_backend() -> str:
     """Run anyio-marked tests on asyncio only."""
     return "asyncio"
-
-
-EngineFactory = Callable[..., Awaitable[VaultEngine]]
 
 
 @pytest.fixture
@@ -55,11 +34,3 @@ def make_engine(tmp_path: Path) -> EngineFactory:
         return engine
 
     return factory
-
-
-def write_raw(engine: VaultEngine, relative: str, text: str) -> Path:
-    """Write a file straight to disk, bypassing the engine (external editor)."""
-    path = engine.root / relative
-    path.parent.mkdir(parents=True, exist_ok=True)
-    path.write_text(text, encoding="utf-8")
-    return path
